+++ src/main/java/com/github/yt/mybatis/query/PageUtils.java
@@ -35,7 +35,7 @@ public class PageUtils {
      * @param <R>        分页数据类型
      * @return 分页信息
      */
-    public static <R> Page createPage(int pageNo, int pageSize, long totalCount, List<R> data) {
+    public static <R> Page createPage(int pageNo, int pageSize, int totalCount, List<R> data) {
         return new Page<R>()
                 .initKey(YtMybatisConfig.pageNoName, YtMybatisConfig.pageSizeName, YtMybatisConfig.pageTotalCountName, YtMybatisConfig.pageDataName)
                 .initValue(pageNo, pageSize, totalCount, data);
